feat(ui,web): move shared primitives into @repowise/ui/shared
Relocates api-error, empty-state, and stat-card from
packages/web/src/components/shared/ into packages/ui/src/shared/ and adds
a ./shared/* subpath export so consumers can import them as
@repowise/ui/shared/<name> (matching the per-file pattern already used).
18 files in packages/web rewrite their @/components/shared/* imports
mechanically. A StatCard smoke test covers label+value, optional
description, and trend rendering. Type-check + production build clean
across all 22 routes; packages/ui test suite is 7/7.
