+++ packages/web/src/app/repos/[id]/search/page.tsx
@@ -5,7 +5,7 @@ import { useParams } from "next/navigation";
 import { Search } from "lucide-react";
 import { SearchBar } from "@/components/search/search-bar";
 import { SearchResultCard } from "@/components/search/search-result-card";
-import { EmptyState } from "@/components/shared/empty-state";
+import { EmptyState } from "@repowise/ui/shared/empty-state";
 import { Skeleton } from "@repowise/ui/ui/skeleton";
 import { useSearch } from "@/lib/hooks/use-search";
