+++ packages/web/src/app/workspace/co-changes/page.tsx
@@ -5,7 +5,7 @@ import { GitMerge, Filter } from "lucide-react";
 import { useWorkspaceCoChanges, useWorkspace } from "@/lib/hooks/use-workspace";
 import { CoChangeTable } from "@/components/workspace/co-change-table";
 import { Card, CardContent, CardHeader, CardTitle } from "@repowise/ui/ui/card";
-import { StatCard } from "@/components/shared/stat-card";
+import { StatCard } from "@repowise/ui/shared/stat-card";
 import { Skeleton } from "@repowise/ui/ui/skeleton";
 
 export default function CoChangesPage() {
