+++ packages/web/src/app/workspace/contracts/page.tsx
@@ -7,7 +7,7 @@ import { useWorkspace } from "@/lib/hooks/use-workspace";
 import { ContractLinksTable } from "@/components/workspace/contract-links-table";
 import { ContractTypeBadge, RoleBadge } from "@/components/workspace/contract-type-badge";
 import { Card, CardContent, CardHeader, CardTitle } from "@repowise/ui/ui/card";
-import { StatCard } from "@/components/shared/stat-card";
+import { StatCard } from "@repowise/ui/shared/stat-card";
 import { Skeleton } from "@repowise/ui/ui/skeleton";
 
 const TYPE_OPTIONS = [
