+++ packages/web/src/components/git/hotspot-table.tsx
@@ -4,7 +4,7 @@ import { useState, useMemo } from "react";
 import { TrendingUp, TrendingDown, Search, Flame, ArrowUpDown, ArrowUp, ArrowDown } from "lucide-react";
 import { Badge } from "@repowise/ui/ui/badge";
 import { Input } from "@repowise/ui/ui/input";
-import { EmptyState } from "@/components/shared/empty-state";
+import { EmptyState } from "@repowise/ui/shared/empty-state";
 import { ChurnBar } from "./churn-bar";
 import { formatLOC } from "@/lib/utils/format";
 import { cn } from "@/lib/utils/cn";
